breakdown home screen to componenets
